rework: Move testsuite file filters and find function to seperate module
